Separate generic transformers methods for STU3 use.
