+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/Transformers.java
@@ -1,0 +1,37 @@
+package gov.va.api.health.dataquery.service.controller;
+
+import java.util.Collection;
+import java.util.Map;
+import lombok.experimental.UtilityClass;
+import org.apache.commons.lang3.StringUtils;
+
+/** Utility methods for transformers that are not specific to one FHIR version. */
+@UtilityClass
+public final class Transformers {
+  /**
+   * Return false if at least one value in the given list is a non-blank string, or a non-null
+   * object.
+   */
+  public static boolean allBlank(Object... values) {
+    for (Object v : values) {
+      if (!isBlank(v)) {
+        return false;
+      }
+    }
+    return true;
+  }
+
+  /** Return true if the value is a blank string, or any other object that is null. */
+  public static boolean isBlank(Object value) {
+    if (value instanceof CharSequence) {
+      return StringUtils.isBlank((CharSequence) value);
+    }
+    if (value instanceof Collection<?>) {
+      return ((Collection<?>) value).isEmpty();
+    }
+    if (value instanceof Map<?, ?>) {
+      return ((Map<?, ?>) value).isEmpty();
+    }
+    return value == null;
+  }
+}
