+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/appointment/AppointmentTransformer.java
@@ -1,6 +1,6 @@
 package gov.va.api.health.dataquery.service.controller.appointment;
 
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.allBlank;
+import static gov.va.api.health.dataquery.service.controller.Transformers.allBlank;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asDateTimeString;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asInteger;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.convert;
