+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/diagnosticreport/DatamartDiagnosticReportTransformer.java
@@ -1,6 +1,6 @@
 package gov.va.api.health.dataquery.service.controller.diagnosticreport;
 
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.allBlank;
+import static gov.va.api.health.dataquery.service.controller.Transformers.allBlank;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.emptyToNull;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.parseInstant;
 import static java.util.Arrays.asList;
@@ -9,6 +9,7 @@ import static org.springframework.util.CollectionUtils.isEmpty;
 
 import gov.va.api.health.argonaut.api.resources.DiagnosticReport;
 import gov.va.api.health.dataquery.service.controller.Dstu2Transformers;
+import gov.va.api.health.dataquery.service.controller.Transformers;
 import gov.va.api.health.dataquery.service.controller.diagnosticreport.DatamartDiagnosticReports.Result;
 import gov.va.api.health.dstu2.api.DataAbsentReason;
 import gov.va.api.health.dstu2.api.datatypes.CodeableConcept;
@@ -34,7 +35,7 @@ final class DatamartDiagnosticReportTransformer {
     if (r == null) {
       return null;
     }
-    if (Dstu2Transformers.isBlank(r.display()) || Dstu2Transformers.isBlank(r.result())) {
+    if (Transformers.isBlank(r.display()) ||Transformers.isBlank(r.result())) {
       return null;
     }
     return Reference.builder().display(r.display()).reference("Observation/" + r.result()).build();
