+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/location/Dstu2LocationTransformer.java
@@ -1,6 +1,6 @@
 package gov.va.api.health.dataquery.service.controller.location;
 
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.allBlank;
+import static gov.va.api.health.dataquery.service.controller.Transformers.allBlank;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asReference;
 import static java.util.Arrays.asList;
 import static org.apache.commons.lang3.StringUtils.isBlank;
