+++ data-query/src/main/java/gov/va/api/health/dataquery/service/controller/observation/ObservationTransformer.java
@@ -1,6 +1,6 @@
 package gov.va.api.health.dataquery.service.controller.observation;
 
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.allBlank;
+import static gov.va.api.health.dataquery.service.controller.Transformers.allBlank;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.asDateTimeString;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.convert;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.convertAll;
