+++ data-query/src/test/java/gov/va/api/health/dataquery/service/controller/Dstu2TransformersTest.java
@@ -14,7 +14,6 @@ import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.e
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.firstPayloadItem;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.hasPayload;
 import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.ifPresent;
-import static gov.va.api.health.dataquery.service.controller.Dstu2Transformers.isBlank;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import gov.va.api.health.dataquery.service.controller.Dstu2Transformers.MissingPayload;
@@ -29,7 +28,6 @@ import java.time.LocalDate;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.function.Function;
 import javax.xml.datatype.DatatypeFactory;
@@ -38,16 +36,6 @@ import lombok.SneakyThrows;
 import org.junit.Test;
 
 public class Dstu2TransformersTest {
-
-  @Test
-  public void allBlank() {
-    assertThat(Dstu2Transformers.allBlank()).isTrue();
-    assertThat(Dstu2Transformers.allBlank(null, null, null, null)).isTrue();
-    assertThat(Dstu2Transformers.allBlank(null, "", " ")).isTrue();
-    assertThat(Dstu2Transformers.allBlank(null, 1, null, null)).isFalse();
-    assertThat(Dstu2Transformers.allBlank(1, "x", "z", 2.0)).isFalse();
-  }
-
   @Test
   public void asCodeableConceptWrappingReturnsNullIfCodingCannotBeConverted() {
     assertThat(asCodeableConceptWrapping(null)).isNull();
@@ -328,17 +316,5 @@ public class Dstu2TransformersTest {
   public void ifPresentReturnsNullWhenObjectIsNull() {
     Function<Object, String> extract = (o) -> "x" + o;
     assertThat(ifPresent(null, extract)).isNull();
-  }
-
-  @Test
-  public void isBlankCollection() {
-    assertThat(isBlank(List.of())).isTrue();
-    assertThat(isBlank(List.of("x"))).isFalse();
-  }
-
-  @Test
-  public void isBlankMap() {
-    assertThat(isBlank(Map.of())).isTrue();
-    assertThat(isBlank(Map.of("x", "y"))).isFalse();
   }
 }
